+++ R/monthly_response.R
@@ -1,25 +1,25 @@
-#' daily_response
+#' monthly_response
 #'
 #' Function calculates all possible values of a selected statistical metric
-#' between one or more response variables and daily sequences of environmental
+#' between one or more response variables and monthly sequences of environmental
 #' data. Calculations are based on moving window which is defined with two
-#' arguments: window width and a location in a matrix of daily sequences of
+#' arguments: window width and a location in a matrix of monthly sequences of
 #' environmental data. Window width could be fixed (use fixed_width) or
 #' variable width (use lower_limit and upper_limit arguments). In this case,
 #' all window widths between lower and upper limit will be used. All calculated
 #' metrics are stored in a matrix. The location of stored calculated metric
 #' in the matrix is indicating a window width (row names) and a location in a
-#' matrix of daily sequences of environmental data (column names).
+#' matrix of monthly sequences of environmental data (column names).
 #'
 #' @param response a data frame with tree-ring proxy variables as columns and
 #' (optional) years as row names. Row.names should be matched with those from a
 #' env_data data frame. If not, set row_names_subset = TRUE.
-#' @param env_data a data frame of daily sequences of environmental data as
-#' columns and years as row names. Each row represents a year and
-#' each column represents a day of a year. Row.names should be matched with
+#' @param env_data a data frame of monthly sequences of environmental
+#' data as columns and years as row names. Each row represents a year and
+#' each column represents a day of a year (or month). Row.names should be matched with
 #' those from a response data frame. If not, set row_names_subset = TRUE.
 #' Alternatively, env_data could be a tidy data with three columns,
-#' i.e. Year, DOY and third column representing values of mean temperatures,
+#' i.e. Year, DOY (Month) and third column representing values of mean temperatures,
 #' sum of precipitation etc. If tidy data is passed to the function, set the argument
 #' tidy_env_data to TRUE.
 #' @param method a character string specifying which method to use. Current
@@ -61,7 +61,7 @@
 #' manually enter the number of components to be used as predictors.
 #' @param eigenvalues_threshold threshold for automatic selection of Principal Components
 #' @param N_components number of Principal Components used as predictors
-#' @param aggregate_function character string specifying how the daily data should be
+#' @param aggregate_function character string specifying how the monthly data should be
 #' aggregated. The default is 'mean', the two other options are 'median' and 'sum'
 #' @param temporal_stability_check character string, specifying, how temporal stability
 #' between the optimal selection and response variable(s) will be analysed. Current
@@ -102,14 +102,14 @@
 #' the highest calculated metric and is obtained by the $plot_extreme output, is the
 #' same for all three reference windows.
 #'
-#' @return a list with 14 elements:
+#' @return a list with 13 elements:
 #' \tabular{rll}{
 #'  1 \tab $calculations   \tab a matrix with calculated metrics\cr
 #'  2 \tab $method \tab the character string of a method \cr
 #'  3 \tab $metric   \tab the character string indicating the metric used for calculations \cr
 #'  4 \tab $analysed_period    \tab the character string specifying the analysed period based on the information from row names. If there are no row names, this argument is given as NA \cr
-#'  5 \tab $optimized_return   \tab data frame with two columns, response variable and aggregated (averaged) daily data that return the optimal results. This data.frame could be directly used to calibrate a model for climate reconstruction \cr
-#'  6 \tab $optimized_return_all    \tab a data frame with aggregated daily data, that returned the optimal result for the entire env_data (and not only subset of analysed years) \cr
+#'  5 \tab $optimized_return   \tab data frame with two columns, response variable and aggregated (averaged) monthly data that return the optimal results. This data.frame could be directly used to calibrate a model for climate reconstruction \cr
+#'  6 \tab $optimized_return_all    \tab a data frame with aggregated monthly data, that returned the optimal result for the entire env_data (and not only subset of analysed years) \cr
 #'  7 \tab $transfer_function    \tab a ggplot object: scatter plot of optimized return and a transfer line of the selected method \cr
 #'  8 \tab $temporal_stability    \tab a data frame with calculations of selected metric for different temporal subsets\cr
 #'  9\tab $cross_validation   \tab a data frame with cross validation results \cr
@@ -133,10 +133,10 @@
 #' data(data_TRW_1)
 #' data(example_proxies_individual)
 #' data(example_proxies_1)
-#' data(LJ_daily_temperatures)
+#' data(LJ_monthly_temperatures)
 #'
 #' # 1 Example with fixed width
-#' example_fixed_width <- daily_response(response = data_MVA, env_data = LJ_daily_temperatures,
+#' example_fixed_width <- monthly_response(response = data_MVA, env_data = LJ_monthly_temperatures,
 #'                                      method = "cor", fixed_width = 0,
 #'                                      row_names_subset = TRUE, remove_insignificant = TRUE,
 #'                                      alpha = 0.05, aggregate_function = 'mean',
@@ -144,13 +144,13 @@
 #' example_fixed_width$plot_extreme
 #'
 #' # 2 Example for past and present
-#' example_MVA_past <- daily_response(response = data_MVA, env_data = LJ_daily_temperatures,
+#' example_MVA_past <- monthly_response(response = data_MVA, env_data = LJ_monthly_temperatures,
 #' method = "cor", lower_limit = 21, upper_limit = 180,
 #' row_names_subset = TRUE, previous_year = TRUE,
 #' remove_insignificant = TRUE, alpha = 0.05,
 #' plot_specific_window = 60, subset_years = c(1940, 1980), aggregate_function = 'sum')
 #'
-#' example_MVA_present <- daily_response(response = data_MVA, env_data = LJ_daily_temperatures,
+#' example_MVA_present <- monthly_response(response = data_MVA, env_data = LJ_monthly_temperatures,
 #'                                       method = "cor", lower_limit = 21, upper_limit = 60,
 #'                                       row_names_subset = TRUE, previous_year = TRUE,
 #'                                       remove_insignificant = TRUE, alpha = 0.05,
@@ -163,8 +163,8 @@
 #' example_MVA_present$plot_specific
 #'
 #' # 3 Example PCA
-#' example_PCA <- daily_response(response = example_proxies_individual,
-#'                               env_data = LJ_daily_temperatures, method = "lm",
+#' example_PCA <- monthly_response(response = example_proxies_individual,
+#'                               env_data = LJ_monthly_temperatures, method = "lm",
 #'                               lower_limit = 21, upper_limit = 180,
 #'                               row_names_subset = TRUE, remove_insignificant = TRUE,
 #'                               alpha = 0.01, PCA_transformation = TRUE,
@@ -174,7 +174,7 @@
 #' example_PCA$plot_heatmap
 #'
 #' # 4 Example negative correlations
-#' example_neg_cor <- daily_response(response = data_TRW_1, env_data = LJ_daily_temperatures,
+#' example_neg_cor <- monthly_response(response = data_TRW_1, env_data = LJ_monthly_temperatures,
 #'                                   method = "cor", lower_limit = 21, upper_limit = 180,
 #'                                   row_names_subset = TRUE, remove_insignificant = TRUE,
 #'                                   alpha = 0.05)
@@ -187,8 +187,8 @@
 #' summary(example_proxies_1)
 #' cor(example_proxies_1)
 #'
-#' example_multiproxy <- daily_response(response = example_proxies_1,
-#'                                      env_data = LJ_daily_temperatures,
+#' example_multiproxy <- monthly_response(response = example_proxies_1,
+#'                                      env_data = LJ_monthly_temperatures,
 #'                                      method = "lm", metric = "adj.r.squared",
 #'                                      lower_limit = 21, upper_limit = 180,
 #'                                      row_names_subset = TRUE, previous_year = FALSE,
@@ -197,7 +197,7 @@
 #' example_multiproxy$plot_heatmap
 #'
 #' # 6 Example to test the temporal stability
-#' example_MVA_ts <- daily_response(response = data_MVA, env_data = LJ_daily_temperatures,
+#' example_MVA_ts <- monthly_response(response = data_MVA, env_data = LJ_monthly_temperatures,
 #' method = "brnn", lower_limit = 100, metric = "adj.r.squared", upper_limit = 180,
 #' row_names_subset = TRUE, remove_insignificant = TRUE, alpha = 0.05,
 #' temporal_stability_check = "running_window", k_running_window = 10)
@@ -206,9 +206,9 @@
 #'
 #' }
 
-daily_response <- function(response, env_data, method = "lm",
-                           metric = "r.squared", lower_limit = 30,
-                           upper_limit = 90, fixed_width = 0,
+monthly_response <- function(response, env_data, method = "lm",
+                           metric = "r.squared", lower_limit = 1,
+                           upper_limit = 12, fixed_width = 0,
                            previous_year = FALSE, neurons = 1,
                            brnn_smooth = TRUE, remove_insignificant = TRUE,
                            alpha = .05, row_names_subset = FALSE,
@@ -246,7 +246,7 @@ daily_response <- function(response, env_data, method = "lm",
  # If there is a column name samp.depth in response data frame, we error msg is displayed
  if ("samp.depth" %in% colnames(response)){
    stop(paste0("response data frame includes samp.depth colname!",
-    " All variables in response data frame will be used as predictors of daily data.",
+    " All variables in response data frame will be used as predictors of monthly data.",
     " Sample depth is not a predictor! Please remove samp.deth column from the response data frame."))
  }
 
@@ -289,7 +289,6 @@ daily_response <- function(response, env_data, method = "lm",
  }
 
 
-
   # PART 1 - general data arrangements, warnings and stops
   # Both bojects (response and env_data) are converted to data frames
   response <- data.frame(response)
@@ -319,11 +318,11 @@ daily_response <- function(response, env_data, method = "lm",
   if (lower_limit >= upper_limit)
     stop("lower_limit can not be higher than upper_limit!")
 
-  if (lower_limit > 365 | lower_limit < 1)
-    stop("lower_limit out of bounds! It should be between 1 and 365")
+  if (lower_limit > 12 | lower_limit < 1)
+    stop("lower_limit out of bounds! It should be between 1 and 12 for monthly response")
 
-  if (upper_limit > 365 | upper_limit < 1)
-    stop("upper_limit out of bounds! It should be between 1 and 365")
+  if (upper_limit > 12 | upper_limit < 1)
+    stop("upper_limit out of bounds! It should be between 1 and 12 for monthly response")
 
 
   # Data manipulation
@@ -388,22 +387,6 @@ daily_response <- function(response, env_data, method = "lm",
                 "At least three characters needed!"))
   }
 
-  # In case of selected window size is less than 14 (2 weeks) or greater than 270 (9 months)
-  if (lower_limit < 14) {
-    warning("Selected lower_limit is less than 14. Consider increasing it!")
-  }
-
-  if (upper_limit > 270) {
-    warning("Selected upper_limit is greater than 270. Consider using lower upper_limit!")
-  }
-
-  if (fixed_width < 14 & fixed_width > 0) {
-    warning("Selected fixed_width is less than 14. Consider increasing it!")
-  }
-
-  if (fixed_width > 270) {
-    warning("Selected fixed_width is greater than 270. Consider using lower fixed_width!")
-  }
 
 
   # If PCA_transformation = TRUE, PCA is performed
@@ -485,7 +468,7 @@ daily_response <- function(response, env_data, method = "lm",
                                                    1 + fixed_width/2 ),0))
       }
 
-      pb <- txtProgressBar(min = 0, max = (ncol(env_data) - fixed_width),
+      pb <- txtProgressBar(min = 0, max = (ncol(env_data) - fixed_width + 1),
                            style = 3)
 
       b = 0
@@ -556,7 +539,7 @@ daily_response <- function(response, env_data, method = "lm",
                                                  1 + fixed_width/2 ),0))
     }
 
-    pb <- txtProgressBar(min = 0, max = (ncol(env_data) - fixed_width),
+    pb <- txtProgressBar(min = 0, max = (ncol(env_data) - fixed_width + 1),
                          style = 3)
 
     b = 0
@@ -630,7 +613,7 @@ daily_response <- function(response, env_data, method = "lm",
                                                  1 + fixed_width/2 ),0))
     }
 
-    pb <- txtProgressBar(min = 0, max = (ncol(env_data) - fixed_width),
+    pb <- txtProgressBar(min = 0, max = (ncol(env_data) - fixed_width + 1),
                          style = 3)
 
     b = 0
@@ -754,11 +737,25 @@ daily_response <- function(response, env_data, method = "lm",
     for (j in 0: (ncol(env_data) - K)) {
 
       if (aggregate_function == 'median'){
-        x <- apply(env_data[1:nrow(env_data), (1 + j) : (j + K)],1 , median, na.rm = TRUE)
+        if (K == 1){
+          x <- env_data[,K+j]
+        } else {
+           x <- apply(env_data[1:nrow(env_data), (1 + j) : (j + K)],1 , median, na.rm = TRUE)}
       } else if (aggregate_function == 'sum'){
+
+        if (K == 1){
+          x <- env_data[,K+j]
+        } else {
+
         x <- apply(env_data[1:nrow(env_data), (1 + j) : (j + K)],1 , sum, na.rm = TRUE)}
+        }
       else if (aggregate_function == 'mean'){
-        x <- rowMeans(env_data[1:nrow(env_data), (1 + j) : (j + K)], na.rm = T)
+
+        if (K == 1){
+          x <- env_data[,K+j]
+        } else {
+
+        x <- rowMeans(env_data[1:nrow(env_data), (1 + j) : (j + K)], na.rm = T)}
       } else {
         stop(paste0("aggregate function is ", aggregate_function, ". Instead it should be mean, median or sum."))
       }
@@ -821,11 +818,25 @@ daily_response <- function(response, env_data, method = "lm",
 
       for (j in 0: (ncol(env_data) - K)) {
         if (aggregate_function == 'median'){
-          x <- apply(env_data[1:nrow(env_data), (1 + j) : (j + K)],1 , median, na.rm = TRUE)
+
+          if (K == 1){
+            x <- env_data[,K+j]
+          } else {
+
+          x <- apply(env_data[1:nrow(env_data), (1 + j) : (j + K)],1 , median, na.rm = TRUE)}
         } else if(aggregate_function == 'sum'){
-          x <- apply(env_data[1:nrow(env_data), (1 + j) : (j + K)],1 , sum, na.rm = TRUE)
+          if (K == 1){
+            x <- env_data[,K+j]
+          } else {
+          x <- apply(env_data[1:nrow(env_data), (1 + j) : (j + K)],1 , sum, na.rm = TRUE)}
         } else if (aggregate_function == 'mean'){
-          x <- rowMeans(env_data[1:nrow(env_data), (1 + j) : (j + K)], na.rm = T)
+
+          if (K == 1){
+            x <- env_data[,K+j]
+          } else {
+
+          x <- rowMeans(env_data[1:nrow(env_data), (1 + j) : (j + K)], na.rm = T)}
+
         } else {
           stop(paste0("aggregate function is ", aggregate_function, ". Instead it should be mean, median or sum."))
         }
@@ -903,11 +914,27 @@ daily_response <- function(response, env_data, method = "lm",
       for (j in 0: (ncol(env_data) - K)) {
 
         if (aggregate_function == 'median'){
-          x <- apply(env_data[1:nrow(env_data), (1 + j) : (j + K)],1 , median, na.rm = TRUE)
+
+          if (K == 1){
+            x <- env_data[,K+j]
+          } else {
+
+          x <- apply(env_data[1:nrow(env_data), (1 + j) : (j + K)],1 , median, na.rm = TRUE)}
         } else if (aggregate_function == 'sum'){
-          x <- apply(env_data[1:nrow(env_data), (1 + j) : (j + K)],1 , sum, na.rm = TRUE)
+
+          if (K == 1){
+            x <- env_data[,K+j]
+          } else {
+
+          x <- apply(env_data[1:nrow(env_data), (1 + j) : (j + K)],1 , sum, na.rm = TRUE)}
+
         } else if (aggregate_function == 'mean'){
-          x <- rowMeans(env_data[1:nrow(env_data), (1 + j) : (j + K)], na.rm = T)
+
+          if (K == 1){
+            x <- env_data[,K+j]
+          } else {
+
+          x <- rowMeans(env_data[1:nrow(env_data), (1 + j) : (j + K)], na.rm = T)}
         } else {
           stop(paste0("aggregate function is ", aggregate_function, ". Instead it should be mean, median or sum."))
         }
@@ -1435,7 +1462,6 @@ daily_response <- function(response, env_data, method = "lm",
     journal_theme +
     ggtitle(paste("Analysed Period:", analysed_period, "\nMethod:", title_string))
 
-analysed_period
   # If there is more than one independent variable in the model,
   # transfer function is not given, since we should return a 3d model
   if (ncol(response) > 1){
@@ -1810,8 +1836,6 @@ for (m in 1:length(empty_list_datasets)){
   }
 
 
-    final_list[[4]]
-
 
     plot_heatmapA <- plot_heatmap(final_list, reference_window = reference_window)
     plot_extremeA <- plot_extreme(final_list, ylimits = ylimits, reference_window = reference_window)
@@ -1847,7 +1871,7 @@ for (m in 1:length(empty_list_datasets)){
                          plot_extreme = plot_extremeA,
                          plot_specific = plot_specificA,
                          PCA_output = PCA_result,
-                         type = "daily")
+                         type = "monthly")
     }
 
     if (method == "cor"){
@@ -1861,7 +1885,7 @@ for (m in 1:length(empty_list_datasets)){
                          plot_extreme = plot_extremeA,
                          plot_specific = plot_specificA,
                          PCA_output = PCA_result,
-                         type = "daily")
+                         type = "monthly")
     }
 
   return(final_list)